Added click add favorite and sign up modal pop up, remove favorite and message to confirm show

+++ js/actions/index.js
@@ -258,32 +258,34 @@ export const getFavoriteError = error => ({
 
 export const deleteFavorites = (product) => dispatch => {
     console.log('product', product)
-    let url = '/favorites';
-    let fetchData = {
-            method: 'DELETE',
-            headers: {
-                'Accept': 'application/json, text/plain, */*',
-                'Content-Type': 'application/json',
-                'Authorization': localStorage.authHeaders,
-            },
-            body: JSON.stringify(product),
-    }
-    return fetch(url, fetchData).then(response => {
-        console.log('response', response.body)
-        if (!response.ok) {
-            throw Error(response.statusText);
+    if (confirm("Are you sure you want to remove the item?")) {
+        let url = '/favorites';
+        let fetchData = {
+                method: 'DELETE',
+                headers: {
+                    'Accept': 'application/json, text/plain, */*',
+                    'Content-Type': 'application/json',
+                    'Authorization': localStorage.authHeaders,
+                },
+                body: JSON.stringify(product),
         }
-        return response;
-    })
-    .then(response => response.json())
-    .then(data => {
-        console.log('data', data);
-        return dispatch(deleteFavoriteSuccess(data));
-    })
-    .catch(error => {
-        console.log('error1', error)
-        return dispatch(deleteFavoriteError(error.message));
-    })
+        return fetch(url, fetchData).then(response => {
+            console.log('response', response.body)
+            if (!response.ok) {
+                throw Error(response.statusText);
+            }
+            return response;
+        })
+        .then(response => response.json())
+        .then(data => {
+            console.log('data', data);
+            return dispatch(deleteFavoriteSuccess(data));
+        })
+        .catch(error => {
+            console.log('error1', error)
+            return dispatch(deleteFavoriteError(error.message));
+        })
+    }
 }
 
 export const DELETE_FAVORITES_SUCCESS = 'DELETE_FAVORITES_SUCCESS';
